move clients to subdir
