+++ WhiteBit.Net/Clients/WhiteBitSocketCommonClient.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace WhiteBit.Net.Clients
+{
+    public class WhiteBitSocketCommonClient
+    {
+        
+    }
+}
